Fix SD Card Diagnostics Logging
This commit fixes an issue where we could not write diagnostics CSV logs to the SD card, because on FAT32 systems, it is apparently prohibited to use an underscore in a directory name.

This commit also moves & renames the consolidated logging/diagnostics logic to the MicroQuad framework so that it can be used in the framework itself.