Improve Debug Logging
This commit adds battery voltage as a column to the diagnostics CSV file that is recorded during a flight.

Also, as far as I can tell, there is literally no way in C/C++ for a macro to forward __VA_ARGS__ to a function AND support an overloaded version without the additional variadic parameters, and so this commit stops using the LOG_INFO macro and instead defines C functions to do this logging.