[Not Working][Firmware] - Set Up Bi-Directional DShot Telemetry
## Context
- I wanted to enable telemetry with each of the Electronic Speed Controllers that run the motors. This would provide valuable motor RPM data, so that I can see how fast each motor is *actually* spinning instead of just what I am commanding it.
- This "DShot" telemetry is bidirectional, on the same wire. My microcontroller transmits a command frame, which takes about 80 microseconds. And then after an 80 microsecond delay, the ESC transmits back a 21 bit telemetry packet.

## Problem
- Since timing is so critical here, I use a higher level "RMT" library to handle transmitting and receiving this data. RMT involves a lot of hardware optimizations for rapid signal RX/TX transmission.
- The ultimate problem is: there literally doesn't seem to be a way to move the GPIO quickly enough from being an RMT TX channel and remap it to be an RMT RX channel.
- I have tried everything I could think of: adding manual blocking microsecond level delays, optimized C callbacks, and so on. And yet fundamentally, the RMT library doesn't seem capable of making this switch quickly enough.
- I will probably revisit this problem in the future.

## Potential Future Solutions
(A) Take a second GPIO, and connect it to the same DShot wire. This way, there is no need to rapidly switch GPIO modes. And then, I can rapidly tell it when to start listening + stop listening for this telemetry.
(B) Try manual bit banging. I've tried this before with the WS2812 LED and could never get the ESP32 timing precise enough, but perhaps there's more that can be done from this angle
(C) Try doing more searching online and on github to see if anyone has finally solved this issue
(D) I never actually whipped out the oscilloscope to verify that the telemetry is even being transmitted by the ESC.