[Firmware] - Improve Calibration
- Previously I had an extremely simplistic calibration model that involved sitting the device flat for several seconds to determine offsets
- This commit implements a new calibration routine that involves tilting the quadcopter through several stages. At each stage, the iOS device then shows the next command (ie. "Place flat", "Roll 90 degrees right", etc)

TODO: In my actual quadcopter, the IMU is oriented 45 degrees offset from the axis of the quadcopter. But this calibration works on the raw IMU data. This means with the current device we need to remove the PCB from the quadcopter frame to perform calibration. I should improve this by doing a misalignment transform during calibration to make this unnecessary